♻️ | AB-7 | Move Avatar.tsx in ui components
